- Updated italian language files. - Fixed a warning in apache plugin.
git-svn-id: svn://svn.ispconfig.org/ispconfig3/branches/ispconfig-3.0.4@2730 c1a6f2c7-8504-0410-933a-a4dae0184b0a

+++ server/plugins-available/apache2_plugin.inc.php
@@ -1757,7 +1757,7 @@ class apache2_plugin {
 			$app->log('Created AWStats config file: '.$awstats_conf_dir.'/awstats.'.$data['new']['domain'].'.conf',LOGLEVEL_DEBUG);
 		}
 		
-		unlink($data['new']['document_root']."/web/stats/index.html");
+		if(is_file($data['new']['document_root']."/web/stats/index.html")) unlink($data['new']['document_root']."/web/stats/index.html");
 		copy("/usr/local/ispconfig/server/conf/awstats_index.php.master",$data['new']['document_root']."/web/stats/index.php");
 	}
